have GCC default to -mfpu=neon-fp16 on webOS
The oldest webOS systems use ARM Cortex-A9 CPUs with NEON and FP16
support. Therefore, it should be safe to enable this for the webOS
toolchain.

Support for NEON/FP16 can be disabled by explicitly passing -mfpu on the
GCC command line. The old behavior can be restored with -mfpu=vfpv3.